+++ src/game/objects/ui/select/index.ts
@@ -1,5 +1,5 @@
-import { ISelectInput } from '../interfaces/selectInput.interface';
-import { Text } from './text';
+import { ISelectInput } from '../../../interfaces/selectInput.interface';
+import { Text } from '../../basic/text';
 
 export class SelectInput extends Phaser.GameObjects.Container {
   label: Text;