+++ src/game/scenes/gameField.ts
@@ -1,7 +1,6 @@
-import { Logo } from '../objects/logo';
-import { HitNote } from '../objects/hitNote';
-import { HitPosition } from '../objects/hitPosition';
-import { NoteAccuracy } from '../objects/noteAccuracy';
+import { HitNote } from '../objects/game/hitNote';
+import { HitPosition } from '../objects/game/hitPosition';
+import { NoteAccuracy } from '../objects/game/noteAccuracy';
 import {
   ENoteAccuracy,
   INotesAccuracyArray,
@@ -11,17 +10,18 @@ import {
   calculateNoteAccuracy,
   calculateOveralAccuracy,
 } from '../core/accuracy';
-import { Text } from '../objects/text';
-import { calculateCurrentScore } from '../core/score';
+import { Text } from '../objects/basic/text';
+import { calculateCurrentScore, Score } from '../core/score';
 import store from '../redux/store';
-import { addHittedNote, getHittedNotes } from '../redux/mapResult';
+import { addHittedNote, getHittedNotes, setCombo } from '../redux/mapResult';
 import { getCurrentMap } from '../redux/currentMap';
 import { IMap } from '../interfaces/map.interface';
 import hitNote from '../../../assets/skin/hitNote.png';
 import hitPosition from '../../../assets/skin/hitPosition.png';
 import { getUserConfig } from '../redux/userConfig';
+import { getCombo } from '../redux/mapResult';
 
-export class MainScene extends Phaser.Scene {
+export class GameField extends Phaser.Scene {
   keyboard: any;
   notesObject: HitNote[] = [];
   scrollSpeed: number = 10;
@@ -32,6 +32,9 @@ export class MainScene extends Phaser.Scene {
   scoreText: any;
   breakAfterLastNote: number = 3000;
   currentMap: IMap;
+  score: Score;
+  comboObject: Text;
+  maxComboObject: Text;
 
   constructor() {
     super({ key: 'MainScene' });
@@ -42,6 +45,7 @@ export class MainScene extends Phaser.Scene {
     this.load.image('hitPosition', hitPosition);
 
     this.currentMap = getCurrentMap();
+    this.score = new Score();
     this.hitPosition = getUserConfig().hitPosition;
   }
 
@@ -74,6 +78,23 @@ export class MainScene extends Phaser.Scene {
       y: 75,
       text: '0',
     });
+    this.comboObject = new Text({
+      scene: this,
+      x: 100,
+      y: 400,
+      text: `combo ${this.score.getCombo().combo}`,
+    });
+    this.maxComboObject = new Text({
+      scene: this,
+      x: 100,
+      y: 450,
+      text: `max combo ${this.score.getCombo().maxCombo}`,
+    });
+  }
+
+  updateScoreUi() {
+    this.comboObject.text = `combo: ${this.score.getCombo().combo}`;
+    this.maxComboObject.text = `max combo: ${this.score.getCombo().maxCombo}`;
   }
 
   createNoteAccuracy(direction: 'up' | 'down', type: ENoteAccuracy) {
@@ -117,6 +138,8 @@ export class MainScene extends Phaser.Scene {
 
               this.createNoteAccuracy('up', accuracy);
               store.dispatch(addHittedNote(accuracy));
+              this.score.increaseCombo();
+              this.updateScoreUi();
             }
             break;
           case 'down':
@@ -124,6 +147,8 @@ export class MainScene extends Phaser.Scene {
               const accuracy = calculateNoteAccuracy(note.delay, time);
               this.createNoteAccuracy('down', accuracy);
               store.dispatch(addHittedNote(accuracy));
+              this.score.increaseCombo();
+              this.updateScoreUi();
             }
             break;
           default:
@@ -135,6 +160,7 @@ export class MainScene extends Phaser.Scene {
       ) {
         this.createNoteAccuracy(note.direction, ENoteAccuracy.Miss);
         store.dispatch(addHittedNote(ENoteAccuracy.Miss));
+        this.score.breakCombo();
       }
     });
   }
@@ -173,6 +199,12 @@ export class MainScene extends Phaser.Scene {
       this.currentMap.notes[this.currentMap.notes.length - 1].delay +
         this.breakAfterLastNote
     ) {
+      store.dispatch(
+        setCombo({
+          combo: this.score.getCombo().combo,
+          maxCombo: this.score.getCombo().maxCombo,
+        }),
+      );
       this.scene.start('ResultScene');
     }
   }