+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.2.0 (2021-04-07)
+
+- create combo
+- create max combo
+- display combo on result screan and in game
+- display max combo on result screan and in game
+
 ## 2.1.0 (2021-04-05)
 
 - create options